making it work for class locally

+++ webwolf/src/main/java/org/owasp/webwolf/FileServer.java
@@ -82,7 +82,7 @@ public class FileServer {
         File[] files = destinationDir.listFiles(File::isFile);
         if (files != null) {
             for (File file : files) {
-                String size = FileUtils.byteCountToDisplaySize(file.length());
+                String size = FileUtils.byteCountToDisplaySize((int) file.length());
                 String link = String.format("files/%s/%s", username, file.getName());
                 uploadedFiles.add(new UploadedFile(file.getName(), size, link));
             }
